+++ src/conversion/importExecution/runConversion.spec.ts
@@ -1,6 +1,6 @@
 import chai, { expect } from "chai";
 import chaiAsPromised from "chai-as-promised";
-import { RunResult as RunResult_V12 } from "../../types/cypress/12.16.0/api";
+import { RunResult as RunResult_V12 } from "../../types/cypress/12.0.0/api";
 import { RunResult as RunResult_V13 } from "../../types/cypress/13.0.0/api";
 import { Status } from "../../types/testStatus";
 import { getTestRunData_V12, getTestRunData_V13 } from "./runConversion";
