+++ src/conversion/importExecution/runConversion.ts
@@ -3,7 +3,7 @@ import {
     RunResult as RunResult_V12,
     ScreenshotInformation as ScreenshotInformation_V12,
     TestResult as TestResult_V12,
-} from "../../types/cypress/12.16.0/api";
+} from "../../types/cypress/12.0.0/api";
 import {
     RunResult as RunResult_V13,
     ScreenshotInformation as ScreenshotInformation_V13,
