+++ src/types/cypress/12.0.0/cypress.ts
